Add i3s 1.6 picking support (#1228)

+++ modules/i3s/src/lib/parsers/constants.js
@@ -5,7 +5,8 @@ export const TYPE_ARRAY_MAP = {
   UInt16: Uint16Array,
   UInt32: Uint32Array,
   Float32: Float32Array,
-  UInt64: Float64Array
+  // eslint-disable-next-line no-undef
+  UInt64: BigUint64Array
 };
 
 export const GL_TYPE_MAP = {
